Replace GetStringAsync with GetByteArrayAsync everywhere
C# will use the webserver's specified encoding, which makes sense, and error without possibility of handling if the encoding is unrecognized, which is a problem. Had code fail because Discord decided to host a file as TIS-620 encoding today. Assume all files are UTF-8 to simplify. (Anything non-UTF8 is wrong and silly and deserves  encoding errors).